+++ src/vector.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vector.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bbouchar <bbouchar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rapelcha <rapelcha@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/10 13:24:58 by bbouchar          #+#    #+#             */
-/*   Updated: 2024/01/17 14:28:48 by bbouchar         ###   ########.fr       */
+/*   Updated: 2024/01/25 08:46:39 by rapelcha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,17 +24,17 @@ t_vector	create_vector(double x, double y, double z)
 	return (vec);
 }
 
-float	length_sqr(t_vector vec)
+double	length_sqr(t_vector vec)
 {
 	return ((vec.x * vec.x) + (vec.y * vec.y) + (vec.z * vec.z));
 }
 
-float	length_vec(t_vector vec)
+double	length_vec(t_vector vec)
 {
 	return (sqrtf((vec.x * vec.x) + (vec.y * vec.y) + (vec.z * vec.z)));
 }
 
-float	dot_vec(t_vector vec1, t_vector vec2)
+double	dot_vec(t_vector vec1, t_vector vec2)
 {
 	return ((vec1.x * vec2.x) + (vec1.y * vec2.y) + (vec1.z * vec2.z));
 }
